splitting the cartouches into own files
