+++ editor/src/components/inspector/sections/component-section/property-control-controls.tsx
@@ -811,187 +811,3 @@ export const Matrix4PropertyControl = React.memo(
     )
   },
 )
-interface DataReferenceCartoucheControlProps {
-  elementPath: ElementPath
-  childOrAttribute: JSXElementChild
-}
-
-export const DataReferenceCartoucheControl = React.memo(
-  (props: DataReferenceCartoucheControlProps) => {
-    const { elementPath, childOrAttribute } = props
-    const dispatch = useDispatch()
-
-    const contentsToDisplay = useEditorState(
-      Substores.metadata,
-      (store) =>
-        getTextContentOfElement(
-          childOrAttribute,
-          MetadataUtils.findElementByElementPath(store.editor.jsxMetadata, elementPath),
-        ),
-      'DataReferenceCartoucheControl contentsToDisplay',
-    )
-
-    const dataPickerButtonData = useDataPickerButton(
-      [EP.parentPath(elementPath)],
-      PP.fromString('children'), // TODO
-      false, // TODO
-      {
-        control: 'none',
-      },
-    )
-
-    const onDeleteCartouche = React.useCallback(() => {}, [])
-
-    return (
-      <>
-        {dataPickerButtonData.popupIsOpen ? dataPickerButtonData.DataPickerComponent : null}
-        <div
-          style={{
-            minWidth: 0, // this ensures that the div can never expand the allocated grid space
-          }}
-          ref={dataPickerButtonData.setReferenceElement}
-        >
-          <IdentifierExpressionCartoucheControl
-            contents={contentsToDisplay.label ?? 'DATA'}
-            icon={contentsToDisplay.type === 'reference' ? '🌸' : <Icons.StringInputControl />}
-            matchType={contentsToDisplay.type === 'reference' ? 'full' : 'none'}
-            onOpenDataPicker={dataPickerButtonData.openPopup}
-            onDeleteCartouche={onDeleteCartouche}
-            safeToDelete={false}
-            testId={'data-reference-cartouche'}
-          />
-        </div>
-      </>
-    )
-  },
-)
-
-interface IdentifierExpressionCartoucheControlProps {
-  contents: string
-  icon: React.ReactChild
-  matchType: 'full' | 'partial' | 'none'
-  onOpenDataPicker: () => void
-  onDeleteCartouche: () => void
-  safeToDelete: boolean
-  testId: string
-}
-export const IdentifierExpressionCartoucheControl = React.memo(
-  (props: IdentifierExpressionCartoucheControlProps) => {
-    const { onDeleteCartouche, testId, safeToDelete } = props
-    const onDelete = React.useCallback<React.MouseEventHandler<HTMLDivElement>>(
-      (e) => {
-        stopPropagation(e)
-        onDeleteCartouche()
-      },
-      [onDeleteCartouche],
-    )
-
-    return (
-      <FlexRow
-        style={{
-          cursor: 'pointer',
-          fontSize: 10,
-          color:
-            props.matchType === 'full'
-              ? colorTheme.white.value
-              : props.matchType === 'partial'
-              ? colorTheme.primary.value
-              : colorTheme.neutralForeground.value,
-          backgroundColor:
-            props.matchType === 'full'
-              ? colorTheme.primary.value
-              : props.matchType === 'partial'
-              ? colorTheme.primary10.value
-              : colorTheme.bg4.value,
-          padding: '0px 4px',
-          borderRadius: 4,
-          height: 22,
-          display: 'flex',
-          flex: 1,
-          gap: 2,
-        }}
-        onClick={props.onOpenDataPicker}
-      >
-        {props.icon}
-        <Tooltip title={props.contents}>
-          <div
-            style={{
-              flex: 1,
-              /* Standard CSS ellipsis */
-              whiteSpace: 'nowrap',
-              overflow: 'hidden',
-              textOverflow: 'ellipsis',
-
-              /* Beginning of string */
-              // direction: 'rtl',
-              textAlign: 'left',
-            }}
-          >
-            {props.contents}
-            &lrm;
-            {/* the &lrm; non-printing character is added to fix the punctuation marks disappearing because of direction: rtl */}
-          </div>
-        </Tooltip>
-        {when(
-          safeToDelete,
-          <Icn
-            category='semantic'
-            type='cross-medium'
-            color='on-highlight-main'
-            width={16}
-            height={16}
-            data-testid={`delete-${testId}`}
-            onClick={onDelete}
-          />,
-        )}
-      </FlexRow>
-    )
-  },
-)
-
-function getTextContentOfElement(
-  element: JSXElementChild,
-  metadata: ElementInstanceMetadata | null,
-): { type: 'literal' | 'reference'; label: string | null } {
-  switch (element.type) {
-    case 'ATTRIBUTE_VALUE':
-      return { type: 'literal', label: `${JSON.stringify(element.value)}` }
-    case 'JSX_TEXT_BLOCK':
-      return { type: 'literal', label: `'${element.text}'` }
-    case 'JS_IDENTIFIER':
-      return { type: 'reference', label: `${element.name}` }
-    case 'JS_ELEMENT_ACCESS':
-      return {
-        type: 'reference',
-        label: `${getTextContentOfElement(element.onValue, null).label}[${
-          getTextContentOfElement(element.element, null).label
-        }]`,
-      }
-    case 'JS_PROPERTY_ACCESS':
-      return {
-        type: 'reference',
-        label: `${getTextContentOfElement(element.onValue, null).label}.${element.property}`,
-      }
-    case 'ATTRIBUTE_FUNCTION_CALL':
-      return { type: 'reference', label: `${element.functionName}(...` }
-    case 'JSX_ELEMENT':
-      return {
-        type: 'literal',
-        label: metadata?.textContent ?? `<${getJSXElementNameLastPart(element.name)}>`,
-      }
-    case 'ATTRIBUTE_NESTED_ARRAY':
-      return { type: 'literal', label: '[...]' }
-    case 'ATTRIBUTE_NESTED_OBJECT':
-      return { type: 'literal', label: '{...}' }
-    case 'JSX_MAP_EXPRESSION':
-      return { type: 'literal', label: 'List' }
-    case 'JSX_CONDITIONAL_EXPRESSION':
-      return { type: 'literal', label: 'Conditional' }
-    case 'ATTRIBUTE_OTHER_JAVASCRIPT':
-      return { type: 'literal', label: element.originalJavascript }
-    case 'JSX_FRAGMENT':
-      return { type: 'literal', label: 'Fragment' }
-    default:
-      assertNever(element)
-  }
-}
